requirements.txt
Flask==3.1.3
python-dotenv==1.0.1
requests==2.31.0
Pillow==10.2.0
google-generativeai==0.5.4
supabase
flask-cors
Flask-Bcrypt==1.0.1
bytez